Extract PageWrapper component to it's file, and make it consistent across the codebase
