+++ src/components/PageWrapper/index.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const PageWrapper = styled.div`
+  max-width: 856px;
+  width: 100%;
+  margin: 0 auto;
+`
