+++ src/pages/LiquidityMining/styleds.ts
@@ -1,13 +1,6 @@
 import styled from 'styled-components'
 import { ButtonPrimary, ButtonSecondary } from '../../components/Button'
-import { AutoColumn } from '../../components/Column'
 import { RowBetween } from '../../components/Row'
-
-export const PageWrapper = styled(AutoColumn)`
-  max-width: 640px;
-  width: 100%;
-  margin-top: -30px;
-`
 
 export const TitleRow = styled(RowBetween)`
   ${({ theme }) => theme.mediaWidth.upToSmall`
