+++ src/pages/Pools/LiquidityMiningCampaign/index.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react'
 import styled from 'styled-components'
 import { Link, Redirect, RouteComponentProps } from 'react-router-dom'
 import { SwapPoolTabs } from '../../../components/NavigationTabs'
-import { PageWrapper } from '../styleds'
+import { PageWrapper } from '../../../components/PageWrapper'
 
 import { TYPE } from '../../../theme'
 import { Box, Flex, Text } from 'rebass'
