+++ src/pages/Pools/styleds.tsx
@@ -1,14 +1,8 @@
 import styled from 'styled-components'
 import { TYPE } from '../../theme'
-import { AutoColumn } from '../../components/Column'
 
 export const Wrapper = styled.div`
   position: relative;
-`
-export const PageWrapper = styled(AutoColumn)`
-  max-width: 832px;
-  width: 100%;
-  margin-top: -30px;
 `
 
 export const ClickableText = styled(TYPE.body)`
